requirements.txt
# RodoviaMonitor Pro - Dependências
requests>=2.31.0,<3.0.0
pyyaml>=6.0,<7.0
openpyxl>=3.1.0,<4.0.0

# Resiliencia
pybreaker>=1.0.1,<2.0.0

# Testes e validação
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
